--- Source Code/.gitattributes
@@ -1,1 +1,0 @@
-*.tar filter=lfs diff=lfs merge=lfs -text
