--- Source Code/clean.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-rm DVRF*.tar.0*
