--- Source Code/merge_and_extract.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-cat DVRF*.tar.0*  > DVRF_v03_source.tar
-tar xzf ./DVRF_v03_source.tar
