+++ Source_Code/merge_and_extract.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+cat DVRF_src_v04*.tar-*  > DVRF_v04.tar.gz
+tar xzf ./DVRF_v04.tar.gz
